moved tests to their folders
